+++ docs/doxygen/doxygen.h
@@ -1,7 +1,7 @@
 /**
 * @file doc/doxygen/doxygen.h
 * @brief Documentation of the main page and the used namespaces.
-* @copyright (c) 2017 Odin Loch Trading as Imortek
+* @copyright (c) 2017 Avast Software, licensed under the MIT license
 */
 
 /**